- add saveArgs - upgrade navigation between fragments - fix SaveCoordinatesDialogFragment.kt when show dialog when not needed - minor fixes - separate xmlRead from xmlWrite
