+++ app/build.gradle
@@ -2,6 +2,7 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-parcelize'
+apply plugin: 'androidx.navigation.safeargs'
 apply plugin: 'dagger.hilt.android.plugin'
 
 android {
@@ -16,6 +17,7 @@ android {
         versionName '6.3.2'
 
         buildConfigField "String", "MY_CLOUD_API", '"http://task.poe.pl.ua:8880/app/"'
+        buildConfigField "String", "MY_PHOTO_STORAGE", '"http://10.107.4.132/abon/"'
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
