+++ app/src/main/java/ua/POE/Task_abon/presentation/model/SearchMap.kt
@@ -1,0 +1,7 @@
+package ua.POE.Task_abon.presentation.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class SearchMap(val map : HashMap<String,String>) : Parcelable
