+++ app/src/main/java/ua/POE/Task_abon/presentation/ui/userinfo/dialog/SaveCoordinatesDialogFragment.kt
@@ -3,6 +3,7 @@ package ua.POE.Task_abon.presentation.ui.userinfo.dialog
 import android.app.Dialog
 import android.content.DialogInterface
 import android.os.Bundle
+import android.util.Log
 import androidx.appcompat.app.AlertDialog
 import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
